update: update new data

+++ tools/download_data.py
@@ -6,6 +6,7 @@ if __name__ == "__main__":
     download_key = "kowndi_pusht_demo_repulse_no_control"
     url_dict = {
         "kowndi_pusht_demo_repulse_no_control": "https://drive.google.com/file/d/1cVdygalbELc7DZYGJB1cu7bbYQ2tueRv/view?usp=sharing",
+        "kowndi_pusht_demo_v0_repulse": "https://drive.google.com/file/d/1XUG6f9qKxwG8lDJ3r64Y36X4L8Lpz-HJ/view?usp=sharing",
     }
     url = url_dict[download_key]
     output = f"{data_dir}/{download_key}.zarr.tar.xz"
@@ -13,7 +14,8 @@ if __name__ == "__main__":
     os.makedirs(data_dir, exist_ok=True)
     if not os.path.exists(output):
         gdown.download(url, output, quiet=False, fuzzy=True)
-        # Untar the file
-        os.system(f"tar -xvf {output} -C {data_dir}")
-        # Remove the tar file
-        os.remove(output)
+        if output.endswith(".tar.xz"):
+            # Untar the file
+            os.system(f"tar -xvf {output} -C {data_dir}")
+            # Remove the tar file
+            os.remove(output)
